Clean up Preview props
